share code and cleanup
